relocate database methods
