+++ src/lib/fetch.js
@@ -1,6 +1,5 @@
 import { _validateEndpoint, optionValidators } from '../validators';
 import { _addQueries, _toArray, _prepareData } from '../utils';
-import { Promise as FirebasePromise } from 'firebase';
 
 export default function _fetch(endpoint, options, db){
   _validateEndpoint(endpoint);
