+++ src/lib/post.js
@@ -1,5 +1,5 @@
 import { database } from 'firebase';
-import { _validateEndpoint, optionValidators } from '../validators';
+import { _validateEndpoint, optionValidators } from './validators';
 
 export default function _post(endpoint, options, db){
   _validateEndpoint(endpoint);
