+++ src/lib/push.js
@@ -1,5 +1,5 @@
 import firebase from 'firebase';
-import { _validateEndpoint, optionValidators } from '../validators';
+import { _validateEndpoint, optionValidators } from './validators';
 
 export default function _push(endpoint, options, db){
   _validateEndpoint(endpoint);
