+++ src/lib/remove.js
